Suggested rework
Alan's initial cleanup suggestions

Hoist onboardingComplete

Hoist onboardingComplete

Fixed spelling of some packages

Fixed spelling of some packages

Moved PEXELS_API_KEY value to gradle.properties

Moved PEXELS_API_KEY value to gradle.properties

Update OnboardingScreens.kt

Added comment to review use of Accompanist Library for Pager & indocators
