+++ .gitignore
@@ -15,3 +15,10 @@
 local.properties
 gradle.properties
 
+.idea/compiler.xml
+.idea/deploymentTargetDropDown.xml
+.idea/inspectionProfiles/Project_Default.xml
+.idea/compiler.xml
+.idea/deploymentTargetDropDown.xml
+.idea/inspectionProfiles/Project_Default.xml
+.idea/deploymentTargetDropDown.xml
