+++ app/build.gradle
@@ -21,10 +21,7 @@ android {
         vectorDrawables {
             useSupportLibrary true
         }
-        buildConfigField("String", "PEXELS_API_KEY", "${PEXELS_API_KEY}")
-
-
-    }
+        buildConfigField("String", "PEXELS_API_KEY", "${PEXELS_API_KEY}")}
 
     buildTypes {
         release {
