--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/comman/data/exceptions/ApiException.kt
@@ -1,3 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.comman.data.exceptions
-
-class ApiException(message: String) : Exception(message)
