--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/comman/data/exceptions/NetworkException.kt
@@ -1,4 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.comman.data.exceptions
-
-
-class NetworkException(message: String, cause: Throwable? = null) : Exception(message, cause)
