--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/comman/di/VideoModule.kt
@@ -1,50 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.comman.di
-
-import com.ameerhamza.animatedgiflivewallpapers.BuildConfig
-import com.ameerhamza.animatedgiflivewallpapers.homePage.data.servies.VideoWallpaperService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object VideoModule {
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor { chain ->
-                val original = chain.request()
-                val requestBuilder = original.newBuilder()
-                    .addHeader(
-                        "Authorization",
-                        BuildConfig.PEXELS_API_KEY
-                    )
-                    .method(original.method(), original.body())
-                val request = requestBuilder.build()
-                chain.proceed(request)
-            }
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(BuildConfig.BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideApiService(retrofit: Retrofit): VideoWallpaperService {
-        return retrofit.create(VideoWallpaperService::class.java)
-    }
-}
