+++ app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/common/data/Result.kt
@@ -1,0 +1,6 @@
+package com.ameerhamza.animatedgiflivewallpapers.common.data
+
+sealed class Result<out T> {
+    data class Success<T>(val data: T) : Result<T>()
+    data class Error(val message: String) : Result<Nothing>()
+}
