+++ app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/common/data/exceptions/ApiException.kt
@@ -1,0 +1,3 @@
+package com.ameerhamza.animatedgiflivewallpapers.common.data.exceptions
+
+class ApiException(message: String) : Exception(message)
