+++ app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/common/data/exceptions/NetworkException.kt
@@ -1,0 +1,4 @@
+package com.ameerhamza.animatedgiflivewallpapers.common.data.exceptions
+
+
+class NetworkException(message: String, cause: Throwable? = null) : Exception(message, cause)
