+++ app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/homePage/ui/HomeScreen.kt
@@ -1,0 +1,22 @@
+package com.ameerhamza.animatedgiflivewallpapers.homePage.ui
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun HomeScreen(viewModel: MainViewModel) {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = "HOME",
+            fontSize = MaterialTheme.typography.h4.fontSize
+        )
+    }
+}
