+++ app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/onboarding/data/dataSource/OnboardingDataSource.kt
@@ -1,0 +1,7 @@
+package com.ameerhamza.animatedgiflivewallpapers.onboarding.data.dataSource
+
+import com.ameerhamza.animatedgiflivewallpapers.onboarding.data.model.OnboardingItem
+
+interface OnboardingDataSource {
+    suspend fun fetchOnboardingItems(): List<OnboardingItem>
+}
