--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/onbording/data/dataSource/OnboardingDataSource.kt
@@ -1,7 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.onbording.data.dataSource
-
-import com.ameerhamza.animatedgiflivewallpapers.onbording.data.model.OnboardingItem
-
-interface OnboardingDataSource {
-    suspend fun fetchOnboardingItems(): List<OnboardingItem>
-}
