--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/onbording/data/model/Settings.kt
@@ -1,11 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.onbording.data.model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "settings")
-data class Settings(
-    @PrimaryKey val id: Int = 0,
-    @ColumnInfo(name = "onboarding_completed") val onboardingCompleted: Boolean = false
-)
