--- app/src/main/java/com/ameerhamza/animatedgiflivewallpapers/onbording/data/repository/OnboardingRepository.kt
@@ -1,15 +1,0 @@
-package com.ameerhamza.animatedgiflivewallpapers.onbording.data.repository
-
-import com.ameerhamza.animatedgiflivewallpapers.comman.data.AppDatabase
-import com.ameerhamza.animatedgiflivewallpapers.onbording.data.model.Settings
-import javax.inject.Inject
-
-class OnboardingRepository @Inject constructor(private val appDatabase: AppDatabase)  {
-     suspend fun insertSettings(settings: Settings) {
-       appDatabase.onBordingDao().insertSettings(settings)
-    }
-
-     suspend fun getSettings(): Settings? {
-        return  appDatabase.onBordingDao().getSettings()
-    }
-}
